export Arrows to an icon component export a color util function
