+++ constants/directions.js
@@ -1,0 +1,6 @@
+export const Directions = {
+    up: 0,
+    down: 180,
+    right: 90,
+    left: 270
+}
